+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -7,6 +7,8 @@ use cmp::Ordering;
 #[cfg(not(feature = "std"))]
 use core::{cmp, convert, fmt, hash, iter, ops, str};
 use hash::Hash;
+#[cfg(feature = "concordium-quickcheck")]
+use quickcheck::Gen;
 #[cfg(feature = "derive-serde")]
 use serde::{Deserialize as SerdeDeserialize, Serialize as SerdeSerialize};
 #[cfg(feature = "derive-serde")]
@@ -39,6 +41,17 @@ pub const ACCOUNT_ADDRESS_SIZE: usize = 32;
 #[cfg_attr(feature = "fuzz", derive(Arbitrary))]
 pub struct Amount {
     pub micro_ccd: u64,
+}
+
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for Amount {
+    fn arbitrary(g: &mut Gen) -> Amount {
+        Amount::from_micro_ccd(quickcheck::Arbitrary::arbitrary(g))
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        Box::new(quickcheck::Arbitrary::shrink(&self.micro_ccd).map(Amount::from_micro_ccd))
+    }
 }
 
 #[cfg(feature = "derive-serde")]
@@ -362,6 +375,19 @@ pub struct Timestamp {
     pub(crate) milliseconds: u64,
 }
 
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for Timestamp {
+    fn arbitrary(g: &mut Gen) -> Timestamp {
+        Timestamp::from_timestamp_millis(quickcheck::Arbitrary::arbitrary(g))
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        Box::new(
+            quickcheck::Arbitrary::shrink(&self.milliseconds).map(Timestamp::from_timestamp_millis),
+        )
+    }
+}
+
 impl Timestamp {
     /// Construct timestamp from milliseconds since unix epoch.
     #[inline(always)]
@@ -627,6 +653,13 @@ impl fmt::Display for Duration {
 #[cfg_attr(feature = "fuzz", derive(Arbitrary))]
 pub struct AccountAddress(pub [u8; ACCOUNT_ADDRESS_SIZE]);
 
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for AccountAddress {
+    fn arbitrary(g: &mut Gen) -> AccountAddress {
+        AccountAddress([0u8; ACCOUNT_ADDRESS_SIZE].map(|_| quickcheck::Arbitrary::arbitrary(g)))
+    }
+}
+
 impl convert::AsRef<[u8; 32]> for AccountAddress {
     fn as_ref(&self) -> &[u8; 32] { &self.0 }
 }
@@ -677,6 +710,31 @@ impl ContractAddress {
     }
 }
 
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for ContractAddress {
+    fn arbitrary(g: &mut Gen) -> ContractAddress {
+        ContractAddress {
+            index:    quickcheck::Arbitrary::arbitrary(g),
+            subindex: quickcheck::Arbitrary::arbitrary(g),
+        }
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        let index = self.index;
+        let subindex = self.subindex;
+        let iter = index
+            .shrink()
+            .map(move |i| {
+                subindex.shrink().map(move |si| ContractAddress {
+                    index:    i,
+                    subindex: si,
+                })
+            })
+            .flatten();
+        Box::new(iter)
+    }
+}
+
 /// Either an address of an account, or contract.
 #[cfg_attr(
     feature = "derive-serde",
@@ -690,6 +748,29 @@ pub enum Address {
     Account(AccountAddress),
     #[cfg_attr(feature = "derive-serde", serde(rename = "AddressContract"))]
     Contract(ContractAddress),
+}
+
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for Address {
+    fn arbitrary(g: &mut Gen) -> Address {
+        //Randomly pick account or contract address.
+        if quickcheck::Arbitrary::arbitrary(g) {
+            Address::Account(quickcheck::Arbitrary::arbitrary(g))
+        } else {
+            Address::Contract(quickcheck::Arbitrary::arbitrary(g))
+        }
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        // Note that shrinking an address does not change its type: account addresses
+        // remain account addresses, the same for contract addresses.
+        match self {
+            Address::Account(a) => Box::new(quickcheck::Arbitrary::shrink(a).map(Address::Account)),
+            Address::Contract(a) => {
+                Box::new(quickcheck::Arbitrary::shrink(a).map(Address::Contract))
+            }
+        }
+    }
 }
 
 impl From<AccountAddress> for Address {
@@ -1215,6 +1296,21 @@ pub struct ChainMetadata {
     pub slot_time: SlotTime,
 }
 
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for ChainMetadata {
+    fn arbitrary(g: &mut Gen) -> ChainMetadata {
+        ChainMetadata {
+            slot_time: quickcheck::Arbitrary::arbitrary(g),
+        }
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        Box::new(quickcheck::Arbitrary::shrink(&self.slot_time).map(|slot_time| ChainMetadata {
+            slot_time,
+        }))
+    }
+}
+
 /// Add offset tracking inside a data structure.
 #[derive(Debug)]
 pub struct Cursor<T> {
@@ -1249,6 +1345,20 @@ impl fmt::Display for NewAttributeValueError {
 #[cfg_attr(feature = "fuzz", derive(Arbitrary))]
 #[derive(Clone, Copy, Debug, Ord, PartialOrd, Eq, PartialEq)]
 pub struct AttributeTag(pub u8);
+
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for AttributeTag {
+    // We choose not to constrain the generated attributes to those currently
+    // defined in `concordium-base/rust-src/id/src/types.rs`. The protocol
+    // supports more attributes and it is reasonable to generate all values
+    // supported by the protocol to ensure that the tested code is robust with
+    // respect to future additions.
+    fn arbitrary(g: &mut Gen) -> AttributeTag { AttributeTag(quickcheck::Arbitrary::arbitrary(g)) }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        Box::new(quickcheck::Arbitrary::shrink(&self.0).map(AttributeTag))
+    }
+}
 
 /// An attribute value.
 /// The meaning of the bytes is dependent on the type of the attribute.
@@ -1287,6 +1397,42 @@ impl AttributeValue {
 
     /// Whether the attribute value has zero length.
     pub fn is_empty(&self) -> bool { self.len() == 0 }
+}
+
+#[cfg(feature = "concordium-quickcheck")]
+fn gen_sized_vec<A: quickcheck::Arbitrary>(g: &mut Gen, size: usize) -> Vec<A> {
+    (0..size).into_iter().map(|_| quickcheck::Arbitrary::arbitrary(g)).collect()
+}
+
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for AttributeValue {
+    fn arbitrary(g: &mut Gen) -> AttributeValue {
+        let size = gen_range_u8(g, 0..32);
+        let mut inner: [u8; 32] = [0u8; 32];
+        let random_data = gen_sized_vec(g, size as usize);
+        inner[1..=size as usize].copy_from_slice(&random_data);
+        inner[0] = size;
+        AttributeValue {
+            inner,
+        }
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        let size = self.inner[0];
+        let data: &[u8] = &self.inner[1..=size as usize];
+        let vs = data.to_vec().shrink();
+        Box::new(
+            vs.map(|v| {
+                let mut inner = [0u8; 32];
+                inner[1..=v.len() as usize].copy_from_slice(&v);
+                inner[0] = v.len() as u8;
+                AttributeValue {
+                    inner,
+                }
+            })
+            .into_iter(),
+        )
+    }
 }
 
 impl AsRef<[u8]> for AttributeValue {
@@ -1404,6 +1550,96 @@ pub struct Policy<Attributes> {
     pub valid_to:          Timestamp,
     /// List of attributes, in ascending order of the tag.
     pub items:             Attributes,
+}
+
+/// Generate a vector of random key-value pairs with no duplication
+/// The length of the resulting vector is <= `size`
+#[cfg(feature = "concordium-quickcheck")]
+fn gen_no_dup_kv_vec<A: quickcheck::Arbitrary + Ord, B: quickcheck::Arbitrary>(
+    g: &mut Gen,
+    size: usize,
+) -> Vec<(A, B)> {
+    use std::collections::BTreeMap;
+    let mut m: BTreeMap<A, B> = BTreeMap::new();
+    for _ in 0..size {
+        let k = A::arbitrary(g);
+        let v = B::arbitrary(g);
+        m.insert(k, v);
+    }
+    m.into_iter().collect()
+}
+
+/// Generate a random `u64` value in the given range by shifting a random `u64`
+/// value using the `%` operator. The reason for doing it this way is that the
+/// range generation method is not exposed by QuickCheck's `Gen`.
+#[cfg(feature = "concordium-quickcheck")]
+fn gen_range_u64(g: &mut Gen, range: core::ops::Range<u64>) -> u64 {
+    let i: u64 = quickcheck::Arbitrary::arbitrary(g);
+    i % (range.end - range.start) + range.start
+}
+
+/// Generate a random `u8` value in the given range by shifting a random `u8`
+/// value using the `%` operator. The reason for doing it this way is that the
+/// range generation method is not exposed by QuickCheck's `Gen`.
+#[cfg(feature = "concordium-quickcheck")]
+fn gen_range_u8(g: &mut Gen, range: core::ops::Range<u8>) -> u8 {
+    let i: u8 = quickcheck::Arbitrary::arbitrary(g);
+    i % (range.end - range.start) + range.start
+}
+
+/// Check that the creation date `created_at` is less then or equal to the
+/// validity date `valid_to`.
+#[cfg(feature = "concordium-quickcheck")]
+fn valid_owned_policy(op: &OwnedPolicy) -> bool {
+    match op {
+        OwnedPolicy {
+            identity_provider: _,
+            created_at,
+            valid_to,
+            items: _,
+        } => created_at <= valid_to,
+    }
+}
+
+#[cfg(feature = "concordium-quickcheck")]
+impl quickcheck::Arbitrary for OwnedPolicy {
+    fn arbitrary(g: &mut Gen) -> OwnedPolicy {
+        let size: u8 = quickcheck::Arbitrary::arbitrary(g);
+        let created_at: Timestamp = quickcheck::Arbitrary::arbitrary(g);
+        // generate `created_at` date so it's <= `valid_to`
+        let valid_to_millis = gen_range_u64(g, created_at.timestamp_millis()..u64::MAX);
+        OwnedPolicy {
+            identity_provider: quickcheck::Arbitrary::arbitrary(g),
+            created_at,
+            valid_to: Timestamp::from_timestamp_millis(valid_to_millis),
+            items: gen_no_dup_kv_vec(g, size as usize),
+        }
+    }
+
+    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
+        let identity_provider = self.identity_provider;
+        let created_at = self.created_at.clone();
+        let valid_to = self.valid_to;
+        let items = self.items.clone();
+        let iter = identity_provider
+            .shrink()
+            .flat_map(move |ip| {
+                let items = items.clone();
+                created_at.shrink().flat_map(move |ca| {
+                    let items = items.clone();
+                    valid_to.shrink().flat_map(move |vt| {
+                        items.shrink().map(move |it| OwnedPolicy {
+                            identity_provider: ip,
+                            created_at:        ca,
+                            valid_to:          vt,
+                            items:             it,
+                        })
+                    })
+                })
+            })
+            .filter(valid_owned_policy);
+        Box::new(iter)
+    }
 }
 
 /// This implementation of deserialize is only useful when used