+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Implement `quickcheck::Arbitrary` for `Timestamp`, `AccountAddress`, `ContractAddress`, `Address`,  `ChainMetadata`, `AttributeTag`, `AttributeValue` and `OwnedPolicy`.
+
 ## concordium-contracts-common 5.0.0 (2022-11-21)
 
 - Add support for smart contract V3 schemas.